Use common push file
