+++ src/push/package.json
@@ -1,6 +1,6 @@
 {
   "name" : "push",
   "main" : "push",
-  "types": "push.d.ts",
+  "types": "push.common.ts",
   "nativescript": {}
 }
