--- src/push/push.d.ts
@@ -1,6 +1,0 @@
-import { EventEmitter } from 'events';
-import { Client } from 'kinvey-js-sdk/dist/client';
-
-export declare class Push extends EventEmitter {
-    constructor(client?: Client);
-}
